+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- support for HCL2 using `python-hcl2` (requires Python >= 3.6)
 
 ## [1.11.0] - 2020-08-11
 ### Added